requirements.txt
docker-py
Flask
Flask-Login
Flask-SQLAlchemy
